Fix issue with underscores in foldcased tags
Vim, when doing foldcase, considers underscore to be "larger" than
alphanumerical characters. However, compareToIgnoreCase on strings seem
to lowercase the chars before comparing. In ASCII, the code for
underscore is larger than upper-cased letters but lower than lowercased
letters: it seems VIM uppercases before comparing while
compareToIgnoreCase lowercases.

This was causing tags files with tags that contain underscores to not be
sorted (according to vim). The fix is simply to upper-case before
comparing.