Add new address form and refactor
Also fix the tests, turns out they were wrong.

This commit makes it so that we use `/search/` instead of `call cursor`.
`call cursor`, thought precise, is brittle. Any change to the file can
cause the address to no longer point to the correct place.

In order to avoid the issue other tag generators have (where the address
points to the beginning of the line) we use `\zs` to point the address
to the actual tag name.